backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.3",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.1",
]
